refactor(planning): extract ProposalStateService from PlanningService
Phase 7 of the cqrs refactor (#227), Task 1.

Lifts the proposal / GitHub-sync / memory-change staging state and its
helpers out of `PlanningService` into a dedicated
`ProposalStateService`. The new service owns:

  - the `proposals` / `gitHubSyncs` / `memoryChanges` Maps and their
    "active" pointers
  - `lastCommitResult` / `lastGitHubSyncResult` / `lastMemoryWriteResult`
  - the issue-id alias resolver and `rewriteProposalIssueAliases`
  - normalize / accumulate / apply-update helpers for the three
    proposal kinds
  - `resetTurnState` / `getStagedWorkItemIds` / `dismissAll`

`PlanningService` shrinks from ~1016 lines to ~604 lines (-412).
Every method that used to read or write the staging state now
delegates to `this.proposalState.X`. Behaviour is unchanged — the
extracted methods retain their bodies verbatim, and the method
calls inside `PlanningService.tools` follow the same call order.

Test harness (`planning.service.test.ts`) constructs a real
`ProposalStateService` instance with the same mocks instead of
passing `{} as never`, so the existing assertions about
issue-id alias resolution and grouped-create staging continue to
exercise the real code paths.

`ProposalStateService` is registered as a provider + export on
`PlanningModule`. Verified via `npm test` (561 / 561) and a clean
dev-server boot returning `{"ok":true}` on `GET /health`.

This is the first half of #227; Task 2 will create the
`PlannerToolRegistry` + extract the 10 individual tool files.
