+++ src/modules/planning/__tests__/planning.service.test.ts
@@ -30,6 +30,7 @@ vi.mock("../../reconciliation/reconciliation.service.js", () => ({
 
 import { checkIdAlignment } from "../../graph/types.js";
 import { PlanningService } from "../planning.service.js";
+import { ProposalStateService } from "../proposal-state.service.js";
 import type { PlanningMutationProposal } from "../types.js";
 
 type CreatedIssue = {
@@ -45,15 +46,25 @@ function makePlanningService(createdIssues: CreatedIssue[]) {
     createIssue.mockResolvedValueOnce(issue);
   }
 
+  const graphService = { getGraph: () => ({ graph_version: "7" }) } as never;
+  const githubService = { createIssue } as never;
+  const memoryService = {} as never;
+  // Phase 7 (#227): proposalState owns the staging/alias state that
+  // PlanningService used to manage inline. Construct a real instance
+  // with the same mocks so the existing assertions about issue-id
+  // alias resolution and grouped staging keep working.
+  const proposalState = new ProposalStateService(graphService, githubService, memoryService);
+
   const service = new PlanningService(
     {} as never,
-    { getGraph: () => ({ graph_version: "7" }) } as never,
+    graphService,
+    {} as never,
+    memoryService,
+    {} as never,
+    githubService,
     {} as never,
     {} as never,
-    {} as never,
-    { createIssue } as never,
-    {} as never,
-    {} as never,
+    proposalState,
   );
 
   (service as any).currentTurnId = "turn_0001";
