+++ src/modules/planning/planning.module.ts
@@ -6,13 +6,14 @@ import { ContextService } from "./context.service.js";
 import { MemoryService } from "./memory.service.js";
 import { PlanningController } from "./planning.controller.js";
 import { PlanningService } from "./planning.service.js";
+import { ProposalStateService } from "./proposal-state.service.js";
 import { SubAgentService } from "./sub-agent.service.js";
 import { ReconciliationModule } from "../reconciliation/reconciliation.module.js";
 
 @Module({
   imports: [GraphModule, GitHubModule, ReconciliationModule, SettingsModule],
   controllers: [PlanningController],
-  providers: [PlanningService, ContextService, MemoryService, SubAgentService],
-  exports: [PlanningService, ContextService, MemoryService, SubAgentService],
+  providers: [PlanningService, ContextService, MemoryService, ProposalStateService, SubAgentService],
+  exports: [PlanningService, ContextService, MemoryService, ProposalStateService, SubAgentService],
 })
 export class PlanningModule {}
